+++ night_light/analyzer/__init__.py
@@ -3,4 +3,4 @@ from .crosswalk_center import *
 from .vehicle_direction import *
 from .distance import *
 from .contrast import *
-from .percieved_brightness import *
+from .brightness import *
